+++ engine/src/Product.java
@@ -1,10 +1,10 @@
 public class Product {
     int reference;
     String name;
-    float price;
+    double price;
     String category;
 
-    public Product(int reference, String name, int price,String category){
+    public Product(int reference, String name, double price,String category){
         this.reference = reference;
         this.name = name;
         this.price = price;
